interface and rename DataSource
