+++ .gitignore
@@ -21,5 +21,6 @@ Thumbs.db
 # Environment files (optional)
 *.env
 
+local.properties
 # Optional: ignore generated files in src (if any)
 # src/**/*.class
