+++ src/main/kotlin/data/source/DataSource.kt
@@ -1,0 +1,7 @@
+package org.baghdad.data.source
+
+import data.source.weather.model.WeatherResponse
+
+interface DataSource {
+    suspend fun fetchCurrentWeather(city: String): WeatherResponse
+}
